Add full deploy and config script
A full deploy and configuration script that can be used to deploy all of
the smart contracts and configure them for the TotemParadox to be
playable has been added.

In addition to the script, the smart contracts were updated where
necessary.

+++ packages/hardhat/contracts/interfaces/IMintingUtils.sol
@@ -6,7 +6,7 @@ import "@openzeppelin/contracts/access/AccessControl.sol";
 error ArrayMismatch();
 
 interface IMintingUtils {
-    event ElementUpdated(string name, uint8 from, uint8 to);
+    event ElementUpdated(string name, uint128 from, uint128 to);
 
     function getRandomElement() external view returns (string memory);
